backend/requirements.txt
fastapi[all]
passlib[bcrypt]
python-jose[cryptography]
sqlalchemy
psycopg2-binary
pydantic-settings